refactor arrow animation into hook
