+++ components/SVG/ArrowLoop/ArrowLoopRight.tsx
@@ -1,5 +1,5 @@
-import { animateWithOptions, drawSVGPath } from 'lib/animations';
-import { RefObject, useRef, useEffect } from 'react';
+import { useArrowAnimation } from 'hooks/useArrowAnimation';
+import { RefObject, useRef } from 'react';
 import styles from './ArrowLoop.module.scss';
 
 type ArrowLoopRightProps = {
@@ -13,59 +13,8 @@ export const ArrowLoopRight = ({
     width = 95,
     height = 123,
 }: ArrowLoopRightProps): React.ReactElement => {
-    const containerRef: RefObject<SVGSVGElement> = useRef<SVGSVGElement | null>(
-        null,
-    );
-
-    // Animate the entrance of the SVG arrow loop.
-    useEffect(() => {
-        const containerNode = containerRef?.current;
-        const hasIOSupport = !!window.IntersectionObserver;
-        if (!hasIOSupport || !containerNode) return;
-
-        const callback = (
-            entries: IntersectionObserverEntry[],
-            observer: IntersectionObserver,
-        ) => {
-            // Look at all intersection observer event entries reported
-            entries.forEach((entry) => {
-                // If the event entry is of a target element intersecting with the observer
-                if (entry.isIntersecting) {
-                    const pathElements = entry.target.children;
-
-                    // Emotion: airy, graceful. Avoid the second path feeling like a tick box check. Draw the user down the page
-                    animateWithOptions(
-                        pathElements[0] as SVGElement,
-                        drawSVGPath,
-                        {
-                            duration: 1700,
-                            easing: 'cubic-bezier(0.45, 0, 0.55, 1)',
-                        },
-                    );
-                    animateWithOptions(
-                        pathElements[1] as SVGElement,
-                        drawSVGPath,
-                        { delay: 1700, duration: 500, easing: 'ease' },
-                    );
-
-                    // Stop observing so the animation doesn't replay if they scroll down again.
-                    observer.unobserve(entry.target);
-                }
-            });
-        };
-
-        let observerThreshold = 0.1;
-
-        const observer = new IntersectionObserver(callback, {
-            root: null,
-            rootMargin: '0px',
-            threshold: observerThreshold,
-        });
-
-        observer.observe(containerNode);
-
-        return () => observer.disconnect();
-    });
+    const svgRef: RefObject<SVGSVGElement> = useRef<SVGSVGElement | null>(null);
+    useArrowAnimation(svgRef);
 
     return (
         <svg
@@ -79,7 +28,7 @@ export const ArrowLoopRight = ({
             strokeLinejoin="round"
             strokeWidth="2.6"
             stroke="#251657"
-            ref={containerRef}
+            ref={svgRef}
         >
             <path
                 className={styles.svgArrowLoopPath}
