+++ components/SVG/ArrowLoop/SmallArrowLoopRight.tsx
@@ -1,21 +1,42 @@
+import { useArrowAnimation } from 'hooks/useArrowAnimation';
+import { RefObject, useRef } from 'react';
+import styles from './ArrowLoop.module.scss';
+
 type SmallArrowLoopRightProps = {
     className?: string;
 };
 
 export const SmallArrowLoopRight = ({
     className = '',
-}: SmallArrowLoopRightProps): React.ReactElement => (
-    <svg
-        xmlns="http://www.w3.org/2000/svg"
-        className={className}
-        width="97"
-        height="129"
-        viewBox="0 0 97 129"
-        fill="none"
-    >
-        <path
-            d="M95.813 35.27a2 2 0 1 0-2.528-3.099l2.528 3.1ZM55.36 56.396l.626 1.9-.626-1.9Zm-29.422-4.73.845-1.813-.845 1.813ZM3.436 12l-1.936-.5 1.936.5Zm5.642-9.265.502 1.936-.502-1.936Zm6.087 3.49-1.29 1.528 1.29-1.528Zm10.457 11.006 1.638-1.147-1.638 1.147ZM36.19 41.626l-1.925.544 1.925-.544Zm2.921 41.142-1.962-.387 1.962.387Zm-24.095 41.681a2 2 0 0 0 1.786 2.193l17.907 1.832a2 2 0 0 0 .407-3.98l-15.917-1.628 1.628-15.917a2 2 0 0 0-3.98-.407l-1.83 17.907Zm78.269-92.278C82.04 41.342 68.413 49.99 54.735 54.496l1.252 3.799C70.29 53.582 84.348 44.622 95.813 35.27l-2.528-3.1Zm-38.55 22.325c-4.261 1.404-9.089 1.27-13.998.24-4.9-1.029-9.743-2.921-13.953-4.884l-1.69 3.625c4.356 2.031 9.508 4.058 14.821 5.173 5.303 1.113 10.903 1.348 16.072-.355l-1.252-3.8Zm-27.951-4.644C11.809 42.87 1.163 28.787 5.372 12.502L1.5 11.5c-4.882 18.886 7.664 34.55 23.593 41.976l1.69-3.625ZM5.372 12.502c.342-1.323.871-3.252 1.682-4.933.87-1.804 1.767-2.7 2.526-2.897L8.575.8C5.97 1.476 4.39 3.885 3.451 5.832 2.453 7.9 1.844 10.168 1.5 11.5l3.872 1Zm4.208-7.83c-.108.028-.025-.035.376.127.373.15.83.42 1.34.788 1.074.772 1.907 1.6 2.579 2.167l2.58-3.056c-.362-.306-1.597-1.476-2.821-2.357-.638-.46-1.387-.93-2.177-1.25-.762-.308-1.798-.572-2.882-.29L9.58 4.67Zm4.294 3.082c3.752 3.168 7.371 6.716 10.11 10.626l3.276-2.295c-3.016-4.306-6.923-8.11-10.805-11.387l-2.58 3.056Zm10.11 10.626c4.918 7.02 7.891 15.35 10.281 23.79l3.85-1.09c-2.424-8.559-5.543-17.413-10.855-24.995l-3.276 2.295Zm10.281 23.79c3.804 13.434 5.574 26.572 2.884 40.21l3.925.775c2.856-14.483.94-28.303-2.96-42.074l-3.849 1.09Zm2.884 40.21c-2.946 14.94-9.683 31.174-21.406 40.722l2.526 3.101C31.034 115.807 38.047 98.5 41.074 83.155l-3.925-.774Z"
-            fill="#251657"
-        />
-    </svg>
-);
+}: SmallArrowLoopRightProps): React.ReactElement => {
+    const svgRef: RefObject<SVGSVGElement> = useRef<SVGSVGElement | null>(null);
+    useArrowAnimation(svgRef);
+
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className={className}
+            width="97"
+            height="129"
+            viewBox="0 0 97 129"
+            fill="none"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth="3.5"
+            stroke="#251657"
+            ref={svgRef}
+        >
+            <path
+                className={styles.svgArrowLoopPath}
+                pathLength="1"
+                d="M94.339 33.6C36.286 87.902-6.275 32.947 3.672 11.748 5.139 8.324 10.473 2.681 15.25 6.204 38.868 23.967 54.746 91.889 18.022 123.778"
+            />
+
+            <path
+                className={styles.svgArrowLoopPath}
+                pathLength="1"
+                d="M18.837 107.144 17.532 124.267 34.818 126.387"
+            />
+        </svg>
+    );
+};
